add toc to docs homepage
